requirements.txt
# Core dependencies for MQGT-SCF empirical validation

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Plotting
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Document processing
python-docx>=1.0.0
pymupdf>=1.23.0
pypdf>=3.15.0

# Web framework (for Zora Brain API)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# HTTP requests (for sensor integration)
requests>=2.31.0

# Optional: For web dashboard
streamlit>=1.28.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
